Unify input styles.
